+++ my-app-frontend/src/Home.js
@@ -1,22 +1,8 @@
 import React from 'react';
-import ActivityItems from './ActivityItems';
 
-function Home({activities}) {
-
-    const mapActivities = activities.map((activity) => {
-        return <ActivityItems
-        id={activity.id}
-        museum={activity.museum}
-        restaurant={activity.restaurant}
-        park={activity.park}
-        misc={activity.misc}
-        />
-    })
-
-
+function Home() {
     return(
     <div className="homeDiv">
-        <ul>{mapActivities}</ul>
         <h1 className="welcome">welcome</h1>
     </div>
     )
